// web/src/types/PropertyTypes.ts
export type PropertyTypes = {
  _id?: string
  _type?: string
  _price?: number
  _size?: number
  _bedrooms?: number
  _rooms?: number
  _bathrooms?: number
  _parking?: number
  _pavement?: number
  _neighborhood?: string
  _city?: string
  _description?: string
  _buy_or_rent?: string
  _cabinets?: boolean
  _intercom?: boolean
  _recreation_area?: boolean
  _private_area?: boolean
  _elevator?: boolean
  _physical_doorman?: boolean
  _electronic_gate?: boolean
  _suite?: number
  _balcony?: boolean
  _photos?: string[]
}